add date calculation, ability to disable chat
The smallest change here is a Boolean argument to the initialize function to enable or disable chat filtering. This is helpful if you only want the BE-GUID

The second part is the function to determine the date of the specific log entry based the current time in the server's timezone. Here's how the logic works:

We know that the server restarts at 0400U (1200Z), so we don't need to worry about multiple days piling up in the same log file (yay). As such we can use 0400U as a trusted delimiter.

If the log line we are parsing is between 0400U and 2359U and the current time is within the same range, the log entry is today. However, if it's currently between 0000U and 0359U we know the log line we are parsing is from yesterday, so we subtract a day.

Alternatively, if the time is between 0000U and 0359U we can easily trust that those times are today. This is due to the 0400U delimiter mentioned above.