So ISO 8601 format is different based on host OS
While integrating this repo with Travis CI my tests started to fail when it came to some of the ISO 8601 business. Turns out, in the infra that Travis CI runs on the iso8601 function was outputting a string, while correct, that was in a different format.

OSX: 2014-01-01T00:00:00Z
TCI: 2014-01-01T00:00:00+00:00

This was breaking tests and causing other random things to fail. This build will fail also but forcibly merge and it should work.