update ServerStats::Base to auto-populate cache
The cache population should happen by default, even with it being a potentially blocking operation. However, because of it being potentially blocking we should allow the ability to turn it off.

This adds an ':auto_cache' option to the ServerStats initializer for turning this feature off.

Add tests to ensure this default behavior is working.