add beguid support to formatters && ruby 2.1.2
This adds the support for printing the BEGUID next to players in the RPT formatter. If you modify the events to contain the :victim_beguid and :offender_beguid keys they will print in the formatters.

This also fixes the name of one test context, and adds build support for Ruby 2.1.2.