remove GameSpy functionality
This removes all GameSpy related functionality from RMuh. As of ArmA 2: Operation Arrowhead 1.63, GameSpy has been removed and the game server no long replies to GameSpy queries. This is pretty fucking lame, if you ask me...

In addition, the examples_spec.rb file was rewritten to no longer require the list of example files. Everything in the examples directory will be tested.

Lastly, the Rubocop maximum LineLength was set to 120. Because ain't nobody got time for that 80 character bullshit... we live in 2014.